Several code simplifications

+++ src/test/scala/com/gonzaloquero/scal8/MemorySpec.scala
@@ -25,20 +25,20 @@ class MemorySpec extends FlatSpec with Matchers {
       memory.get(i) should be(UByte.of(0))
     }
 
-    memory.get(0x200) should be(UByte.of(0x00.toByte))
-    memory.get(0x201) should be(UByte.of(0xE0.toByte))
-    memory.get(0x202) should be(UByte.of(0xA2.toByte))
-    memory.get(0x203) should be(UByte.of(0x48.toByte))
+    memory.get(0x200) should be(UByte.of(0x00))
+    memory.get(0x201) should be(UByte.of(0xE0))
+    memory.get(0x202) should be(UByte.of(0xA2))
+    memory.get(0x203) should be(UByte.of(0x48))
   }
 
   "get" should "return the value in the correspondent memory address" in {
     val memory = Memory.withProgram(Array[UByte](0x48))
-    memory.get(0x200) should be(UByte.of(0x48.toByte))
+    memory.get(0x200) should be(UByte.of(0x48))
   }
 
   "set" should "change the value of the memory address" in {
     val memory = Memory.withProgram(Array[UByte]())
     memory.set(0x200, 0x48)
-    memory.get(0x200) should be(UByte.of(0x48.toByte))
+    memory.get(0x200) should be(UByte.of(0x48))
   }
 }
